Реорганизация проекта. Доработан класс работы с файлом и списком вакансий, появилась возможность вывода вакансии по индексу, внесение в список вакансий как в виде словаря так и в виде объекта, можно добавлять списки словарей или объектов. Возможность вывода списка вакансий в str виде либо расширенном виде.
