+++ README.MD
@@ -3,7 +3,8 @@
 Разрабатывается функционал приложения по работе с вакансиями.
 Разработаны классы:
 - описания и сравнения вакансий по зараплате;
-- по записи/считыванию/добавлению/убавлению вакансий;
+- по созданию и редактированию списка вакансий;
+- по чтению и записи файла с вакансиями;
 
 ## Установка:
 1. Клонируйте репозиторий:
