+++ src/vacansy.py
@@ -1,0 +1,47 @@
+from functools import total_ordering
+
+
+@total_ordering
+class Vacansy:
+    """Класс, описывающий вакансию и её основные характеристики"""
+
+    title: str
+    link: str
+    salary: int | float
+    description: str
+    requirement: str
+
+    def __init__(self, title, link, description, requirement, salary=0):
+        self.title = title
+        self.link = link
+        self.salary = salary
+        self.description = description
+        self.requirement = requirement
+
+    def __str__(self):
+        return f"Вакансия - {self.title}, зарплата - {self.salary}, ссылка на вакансию - {self.link}."
+
+    def __eq__(self, other):
+        if isinstance(other, Vacansy):
+            return self.salary == other.salary
+        else:
+            return NotImplemented
+
+    def __lt__(self, other):
+        if isinstance(other, Vacansy):
+            return self.salary < other.salary
+        else:
+            return NotImplemented
+
+    def __ge__(self, other):
+        if isinstance(other, Vacansy):
+            return self.salary >= other.salary
+        else:
+            return NotImplemented
+
+
+if __name__ == "__main__":
+    vac_1 = Vacansy("Продавец", "link.ru", "Продажа томатов.", "Без опыта работы.", 30000)
+    print(vac_1.description)
+    print(vac_1.requirement)
+    print(vac_1)
