+++ tests/conftest.py
@@ -1,6 +1,26 @@
 import pytest
 
-from src.classes import Vacansy
+from src.vacansy import Vacansy
+
+
+@pytest.fixture
+def vacansy_list():
+    return [
+        Vacansy(
+            title="Инженер 1кат",
+            link="artemtim.ru",
+            salary=90000,
+            description="Работа с технической документацией",
+            requirement="Опрыт работы от 3 лет. Высшее образование.",
+        ),
+        Vacansy(
+            title="Инженер",
+            link="artemtim.ru",
+            salary=50000,
+            description="Работа с технической документацией",
+            requirement="Опрыт работы от 3 лет. Высшее образование.",
+        ),
+    ]
 
 
 @pytest.fixture
@@ -12,6 +32,7 @@ def vacansy_1():
         description="Работа с технической документацией",
         requirement="Опрыт работы от 3 лет. Высшее образование.",
     )
+
 
 @pytest.fixture
 def vacansy_2():
