+++ tests/test_vacansy.py
@@ -1,6 +1,6 @@
 import pytest
 
-from src.classes import Vacansy
+from src.vacansy import Vacansy
 
 
 # Тест инициализации класса Vacansy
@@ -10,7 +10,8 @@ def test_vacansy_init(vacansy_1):
     assert vacansy_1.salary == 50000
     assert vacansy_1.description == "Работа с технической документацией"
     assert vacansy_1.requirement == "Опрыт работы от 3 лет. Высшее образование."
-    assert str(vacansy_1) == 'Вакансия - Инженер, зарплата - 50000, ссылка на вакансию - artemtim.ru.'
+    assert str(vacansy_1) == "Вакансия - Инженер, зарплата - 50000, ссылка на вакансию - artemtim.ru."
+
 
 def test_vacansy_comparision(vacansy_1, vacansy_2):
     assert vacansy_1 < vacansy_2
